Move 'core' models to their own folder
