+++ rgd/stac/models/extensions/core/models/__init__.py
@@ -1,0 +1,7 @@
+"""New models created by the 'core' extension."""
+
+from .asset_role import AssetRole
+from .collection_keyword import CollectionKeyword
+from .media_type import MediaType
+
+__all__ = ['AssetRole', 'CollectionKeyword', 'MediaType']
